backend/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
PyJWT==2.8.0
timedelta==2020.12.3